Optimization, Crash logging, APIs, Profiling, GLM Several `new` operator calls have been replaced with stack allocation, primarily for `Gpu*` classes. This has meant that they no longer use RAII and instead have dedicated `::Initialize` and `::Delete` functions
Fixed logs not being saved if the engine caught an exception and showed the crash message

Removed the `SpotLight.InnerCone` and `OuterCone` properties and replaced them with `SpotLight.Angle`

Added new Profiler scopes for Event Callbacks, Shadow rendering and Post processing

GLM is now a submodule

Hopefully fixed that bug on Linux where it would log a warning for generating mipmaps on a 0x0 texture

Attempts at implementing Normal Maps didn't quite work right, there's just dead code left over in the shaders now

Fixed a compiler warning for the `Renderer` class, as well as one in the Model Importer, and made several GPU Buffer functions `const`

+++ src/impl/asset/ModelImporter.cpp
@@ -375,7 +375,7 @@ void ModelLoader::m_LoadMesh(uint32_t MeshIndex)
 	for (const glm::vec3& position : positions)
 		center += position;
 
-	center /= positions.size();
+	center /= static_cast<float>(positions.size());
 
 	for (size_t index = 0; index < positions.size(); index++)
 		positions[index] -= center;
